% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{abline,RobReg-method}
\alias{abline,RobReg-method}
\title{Add line to plot}
\usage{
\S4method{abline}{RobReg}(a = NULL, b = NULL, h = NULL, v = NULL,
  reg = NULL, coef = NULL, untf = FALSE, ...)
}
\arguments{
\item{a}{an object of class RobReg}

\item{b}{not used.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}

\item{h}{the y-value(s) for horizontal line(s).}

\item{v}{the x-value(s) for vertical line(s).}

\item{reg}{an object with a \code{\link{coef}} method.  See \sQuote{Details}.}

\item{coef}{a vector of length two giving the intercept and slope.}

\item{untf}{logical asking whether to \emph{untransform}.  See
    \sQuote{Details}.}
}
\description{
Add fitted line to a plot. This is overloaded function for robust regression methods from package depthproc.
}

