% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{abline_depthproc}
\alias{abline_depthproc}
\title{Add line to plot}
\usage{
\S4method{abline}{RobReg}(a = NULL, b = NULL, h = NULL, v = NULL,
  reg = NULL, coef = NULL, untf = FALSE, ...)
}
\arguments{
\item{a}{the intercept and slope, single values.}

\item{h}{the y-value(s) for horizontal line(s).}

\item{v}{the x-value(s) for vertical line(s).}

\item{reg}{an object with a \code{\link{coef}} method.  See \sQuote{Details}.}

\item{coef}{a vector of length two giving the intercept and slope.}

\item{untf}{logical asking whether to \emph{untransform}.  See
    \sQuote{Details}.}

\item{...}{\link{graphical parameters} such as
    \code{col}, \code{lty} and  \code{lwd} (possibly as vectors: see
    \sQuote{Details}) and \code{xpd} and the line characteristics
    \code{lend}, \code{ljoin} and \code{lmitre}.}
}
\description{
Add fitted line to a plot. This is overloaded function for robust regression methods from package depthproc.
}

