% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{addLSDContour,LSDepthContour-method}
\alias{addLSDContour,LSDepthContour-method}
\title{Add location-scale depth contour to a plot}
\usage{
\S4method{addLSDContour}{LSDepthContour}(x, cont = NULL, ...)
}
\description{
Add location-scale depth contour to a plot
}

