\docType{class}
\name{BinnDepth2d-class}
\alias{BinnDepth2d-class}
\title{The title for my S4 class that extends \code{"character"} class.}
\description{
Binn
}
\details{
\describe{ \item{myslot1}{A logical keeping track of
something.}

\item{myslot2}{An integer specifying something else.}

\item{myslot3}{A data.frame holding some data.} }
}

