\docType{class}
\name{BinnDepth2d}
\alias{BinnDepth2d}
\title{The title for my S4 class that extends \code{"character"} class.}
\description{
Some details about this class and my plans for it in the
body.
}
\section{Slots}{
  \describe{ \item{freq}{Matrix with number of elements in
  certain bin.} \item{mid_x}{Middle values on x-axis.}
  \item{mid_y}{Middle values on y-axis.}
  \item{breaks_x}{Boundaries of bins.}
  \item{breaks_y}{Boundaries of bins.}
  \item{input_data}{Binned data.} \item{max_depth_x}{Point
  with maximum depth on x-axis.} \item{max_depth_y}{Point
  with maximum depth on y-axis.} }
}

\section{Slots}{
  \describe{ \item{freq}{Matrix with number of elements in
  certain bin.} \item{mid_x}{Middle values on x-axis.}
  \item{mid_y}{Middle values on y-axis.}
  \item{breaks_x}{Boundaries of bins.}
  \item{breaks_y}{Boundaries of bins.}
  \item{input_data}{Binned data.} \item{max_depth_x}{Point
  with maximum depth on x-axis.} \item{max_depth_y}{Point
  with maximum depth on y-axis.} }
}

