% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DepthCurveClass.R
\docType{methods}
\name{combineDepthCurves}
\alias{combineDepthCurves}
\alias{combineDepthCurves,ANY,ANY,list-method}
\alias{combineDepthCurves,DepthCurveList,DepthCurve,ANY-method}
\alias{combineDepthCurves,DepthCurve,DepthCurveList,ANY-method}
\alias{combineDepthCurves,DepthCurve,DepthCurve,ANY-method}
\title{Adds plots}
\usage{
combineDepthCurves(x, y, .list = NULL)

\S4method{combineDepthCurves}{ANY,ANY,list}(x, y, .list = NULL)

\S4method{combineDepthCurves}{DepthCurveList,DepthCurve,ANY}(x, y, .list = NULL)

\S4method{combineDepthCurves}{DepthCurve,DepthCurveList,ANY}(x, y, .list = NULL)

\S4method{combineDepthCurves}{DepthCurve,DepthCurve,ANY}(x, y, .list = NULL)
}
\arguments{
\item{x}{object}

\item{y}{object}

\item{.list}{list of plots to combine.}
}
\description{
Adds plots
}
\details{
See \code{\link{DepthCurve-class}} for description.
}
