% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceFunctions.R
\docType{class}
\name{CovDepthWeighted-class}
\alias{CovDepthWeighted-class}
\title{CovLP}
\description{
This class, derived from the virtual class "CovRobust" accomodates weighted by \eqn{ L ^ p } depth multivariate location and scatter estimator.
}
\details{
See \code{\link{CovLP}} for the function used to calculate weighted by \eqn{ L ^ p } depth covariance matrix.
}
