% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthproc.R
\docType{data}
\name{cracow.airpollution}
\alias{cracow.airpollution}
\title{Air pollution with PM10 in Cracow within day and night in December 2016}
\format{data frame containing 744 rows.}
\usage{
data("cracow.airpollution")
}
\description{
Air pollution with PM10 in Cracow within day and night in December 2016
}
\references{
1. Kosiorowski D, Rydlewski J P, Zawadzki Z (2017). 'Functional Outliers Detection By The Example Of Air Quality Monitoring.' submitted.

2. Kosiorowski D, Szlachtowska E (2017). 'K- local Median Algorithm for Functional Data in Empirical Analysis of Air Pollution Data.' Proceedings from the 11th Professor A. Zelias International Conference, pp. 153–162. Cracow University of Economics.
}
\keyword{datasets}
