% Generated by roxygen2 (4.0.1): do not edit by hand
\name{depthDensity}
\alias{depthDensity}
\title{Depth weighted density estimator}
\usage{
depthDensity(x, y, nx = 5, ny = 32, xg = NULL, yg = NULL, ...)
}
\description{
Depth weighted density estimator
}
\examples{
data = depthproc:::.sampleData(1:5, 100)
x = data[,1]; y = data[,2]
plot(x,y)
dep = depthDensity(x,y)
plot(dep, type = "raw")
plot(dep, type = "depth")
}

