% Generated by roxygen2 (4.0.2): do not edit by hand
\name{depthMBD}
\alias{depthMBD}
\title{Modified band depth}
\usage{
depthMBD(u, X, name = "X", ...)
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).}
}
\description{
Computes the modified band depth.
}
\examples{
x = matrix(rnorm(600), nc = 20)
 depthMBD(x)
}

