% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fncBoxPlot}
\alias{fncBoxPlot}
\title{Functional boxplot based on Modified Band Depth}
\usage{
fncBoxPlot(x, bands = c(0, 0.5), method = "MBD", type = "ggplot2", ...)
}
\arguments{
\item{x}{data matrix}

\item{bands}{limits for bands}

\item{method}{depth method}

\item{type}{method used for create plot ("ggplot2" or "base")}

\item{\dots}{other arguments passed to fncDepth}
}
\description{
Functional boxplot based on Modified Band Depth
}
\examples{
x  = matrix(rnorm(200), ncol = 10)
fncBoxPlot(x, bands = c(0, 0.5, 1), method = "FM")
fncBoxPlot(x, bands = c(0, 0.5, 1), method = "MBD")

# fncBoxPlot handles zoo and xts objects
library(xts)
x  = matrix(rnorm(200), ncol = 10)
time = as.POSIXct(1:nrow(x) * 86400, origin = "1970-01-01")
x = xts(x, order.by = time)
fncBoxPlot(x, bands = c(0, 0.5, 1), method = "FM")
}

