% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_depths.R
\name{fncDepthBD}
\alias{fncDepthBD}
\title{Band Depth}
\usage{
fncDepthBD(u, X)
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data according to which the depth is calculated, optional}
}
\description{
Computes the band depth.
}
\examples{

x <- matrix(rnorm(60), nc = 20)
DepthProc::fncDepthMBD(x)
DepthProc::fncDepthMBD(x, x)

}
