% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_depths.R
\name{fncDepthFM}
\alias{fncDepthFM}
\title{FM Depth}
\usage{
fncDepthFM(u, X, dep1d_params = list(method = "Projection"))
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data as a matrix. If it is a matrix or data frame, then each row is viewed as one multivariate observation.}

\item{dep1d_params}{parameters passed to depth function used in one dimension.}
}
\description{
Computes Frainman-Muniz depth for functional data.
}
\examples{
x <- matrix(rnorm(60), nc = 20)
DepthProc::fncDepthFM(x)

}
