% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_depths.R
\name{fncDepthMBD}
\alias{fncDepthMBD}
\title{Modified band depth}
\usage{
fncDepthMBD(u, X)
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data as a matrix. If it is a matrix or data frame, then each row is viewed as one multivariate observation.}
}
\description{
Computes the modified band depth.
}
\examples{

x <- matrix(rnorm(60), nc = 20)
fncDepthMBD(x)
fncDepthMBD(x, x)

}
