% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncFunctions.R
\name{fncDepthMedian}
\alias{fncDepthMedian}
\title{Functional median}
\usage{
fncDepthMedian(u, X = NULL, method = "MBD", byrow = NULL, unique = TRUE, ...)
}
\arguments{
\item{u}{data matrix}

\item{X}{reference set. If null u will be used as reference.}

\item{method}{depth method}

\item{byrow}{byrow}

\item{unique}{if true}

\item{\dots}{other arguments passed to fncDepth}
}
\description{
Calculate functional median based on data depth.
}
\examples{

x <- matrix(rnorm(600), nc = 20)
md <- DepthProc::fncDepthMedian(x, method = "FM", dep1d_params = list(method = "Mahalanobis"))

}
