% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthproc.R
\docType{data}
\name{france}
\alias{france}
\title{Relation between minimum wage (MW) and unemployment rate (UR) in France.}
\format{
data frame containing 17 rows and two column. MW is a minimum wage, and UR is an unemployment rate.
}
\usage{
data(france)
}
\description{
Relation between minimum wage (MW) and unemployment rate (UR) in France.
}
\keyword{datasets}
