% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassesFnc.R
\docType{class}
\name{FunctionalDepth-class}
\alias{FunctionalDepth-class}
\title{Functional Depth}
\description{
Virtual class with structure for every functional depth class from depthproc package. Inherits from \code{\link{Depth-class}}.
}
\section{Slots}{

\describe{
\item{\code{index}}{numeric, or time-based object.}
}}

