% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getLSDContour,LSDepthContour-method}
\alias{getLSDContour,LSDepthContour-method}
\title{Get location scale contour from LSDepthContour object}
\usage{
\S4method{getLSDContour}{LSDepthContour}(x, cont, ...)
}
\description{
Get location scale contour from LSDepthContour object
}

