% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{combineDepthCurves}
\alias{combineDepthCurves}
\title{Adds plots}
\usage{
combineDepthCurves(x, y, .list = NULL)
}
\arguments{
\item{x}{object}

\item{y}{object}

\item{.list}{list of plots to combine.}
}
\description{
Adds plots
}
\details{
See \code{\link{DepthCurve-class}} for description.
}
