% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{inf.mort}
\alias{inf.mort}
\title{Infant mortality rate (0-1 year) per 1,000 live births}
\format{A data frame with 654 rows and 4 variables}
\source{
http://mdgs.un.org/unsd/mdg/Data.aspx
}
\usage{
data(inf.mort)
}
\description{
Infant mortality rate (0-1 year) per 1,000 live births
}
\keyword{datasets}

