% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthproc.R
\docType{data}
\name{internet.users}
\alias{internet.users}
\title{Internet view data}
\format{data frame containing 17518 rows and 6 columns - 17518 working days of the 
Internet service considered with respect to variables:
service, month, day, hour, unique users and page views.}
\usage{
data(internet.users)
}
\description{
Internet view data
}
\references{
Kosiorowski, Rydlewski, Snarska (2016), Detecting a Structural Change in Functional Time Series Using Local Wilcoxon Statistic, arXiv: 1604.03776v2
}
\keyword{datasets}
