% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthproc.R
\docType{data}
\name{katowice.airpollution}
\alias{katowice.airpollution}
\title{Air pollution in Katowice city by hour.}
\format{data frame containing 181 rows (days) and 24 columns. Each column is an air pollination for given hour.}
\usage{
data("katowice.airpollution")
}
\description{
Air pollution in Katowice city by hour.
}
\keyword{datasets}
