% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{listClass}
\alias{listClass}
\alias{listClass,DepthCurve-method}
\alias{listClass,ScaleCurve-method}
\alias{listClass,AsymmetryCurve-method}
\title{Container class for a DepthCurve}
\usage{
listClass(object)

\S4method{listClass}{DepthCurve}(object)

\S4method{listClass}{ScaleCurve}(object)

\S4method{listClass}{AsymmetryCurve}(object)
}
\arguments{
\item{object}{an object that inherits from \link{DepthCurve-class}.}
}
\description{
Returns the name of the \link{DepthCurveList-class} class that holds curves of
\code{object}'s class. \code{plot()} on a single curve and
\code{\link{combineDepthCurves}} on two of them both need that name, and
every \code{DepthCurve} subclass declares it with its own method rather than
having it inferred from the subclass name.
}
