% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationScaleDepth.R
\docType{methods}
\name{lsdAddContour}
\alias{lsdAddContour}
\alias{lsdAddContour,LSDepthContour-method}
\title{Adds location scale depth contour to the existing plot.}
\usage{
lsdAddContour(x, cont = NULL, ...)

\S4method{lsdAddContour}{LSDepthContour}(x, cont = NULL, ...)
}
\arguments{
\item{x}{object of class LSDepthContour}

\item{cont}{depth of contour to plot}

\item{...}{other arguments passed to polygon function}
}
\description{
This function add one location-scale contour to the existing plot.
}
\examples{
smp <- rf(100, 5, 10)
x <- lsdSampleDepthContours(smp)
plot(x)
lsdAddContour(x, 0.1, col = "grey50")
lsdAddContour(x, 0.3, col = "grey10", border = "red", lwd = 4)
}
