% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lsdAddContour}
\alias{lsdAddContour}
\title{Generic function for lsdAddContour}
\usage{
lsdAddContour(x, cont = NULL, ...)
}
\arguments{
\item{x}{object of class LSDepthContour}

\item{cont}{depth of contour to plot}

\item{...}{other arguments passed to polygon function}
}
\description{
Generic function for lsdAddContour
}
\examples{
smp = rf(100,5,10)
x = lsdSampleDepthContours(smp)
plot(x)
lsdAddContour(x,0.1, col = "grey50")
lsdAddContour(x,0.3, col = "grey10", border = "red", lwd = 4)
}

