% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationScaleDepth.R
\docType{class}
\name{LSDepth-class}
\alias{LSDepth-class}
\title{Location-Scale depth class}
\description{
Class used to store maximum location-scale depth results.
}
\section{Slots}{

\describe{
\item{\code{max_depth}}{maximum Student depth value.}

\item{\code{mu}}{location estimate in the deepest point.}

\item{\code{sigma}}{scale estimate in the deepest point.}
}}

