% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lsdSampleMaxDepth}
\alias{lsdSampleMaxDepth}
\title{Calculates the maximum sample location-scale depth}
\usage{
lsdSampleMaxDepth(x, iter = 100, eps = 1e-04, p_length = 10)
}
\arguments{
\item{x}{one dimensional vector with sample}

\item{iter}{maximum number of iterations in algorith for calculation Location-Scale Depth}

\item{eps}{tolerance level}

\item{p_length}{is the maximum length of the precision step at the end}
}
\description{
Calculates the maximum sample location-scale depth
}
\examples{
x = rnorm(100)
lsdSampleMaxDepth(x)
y = rf(100, 4,10)
lsdSampleMaxDepth(y)
}

