% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{maesles.imm}
\alias{maesles.imm}
\title{Children 1 year old immunized against measles, percentage}
\format{A data frame with 654 rows and 4 variables}
\source{
http://mdgs.un.org/unsd/mdg/Data.aspx
}
\usage{
data(maesles.imm)
}
\description{
Children 1 year old immunized against measles, percentage
}
\keyword{datasets}

