% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maxSampleLocScaleDepth}
\alias{maxSampleLocScaleDepth}
\title{Calculates the maximum sample location-scale depth}
\usage{
maxSampleLocScaleDepth(x, iter = 100, eps = 1e-04, p_length = 10)
}
\description{
Calculates the maximum sample location-scale depth
}

