% Generated by roxygen2 (4.0.1): do not edit by hand
\name{medianDepthConfInterval}
\alias{medianDepthConfInterval}
\title{Bootstrap confidence interval for depth median}
\usage{
medianDepthConfInterval(x, method = "Projection", R = 1000, ...)
}
\description{
Bootstrap confidence interval for depth median
}

