% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,BinnDepth2d,ANY-method}
\alias{plot,BinnDepth2d,ANY-method}
\title{2d Binning plot}
\usage{
\S4method{plot}{BinnDepth2d,ANY}(x, ..., alpha = 0.1, bg_col = "red",
  add_mid = TRUE)
}
\arguments{
\item{x}{object of class BinnDepth2d}

\item{\dots}{graphical parameters passed to plot}

\item{alpha}{alpha value for rgb function}

\item{bg_col}{backgroud color}

\item{add_mid}{logical. If TRUE centers of binns will be marked.}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\description{
Binning 2d
}
\examples{
require(MASS)
 tmp = binningDepth2D(x = mvrnorm(100,rep(0,2),diag(2)))
 plot(tmp)
}
\seealso{
\code{\link{depth}}
}
\keyword{depth}
\keyword{function}
\keyword{multivariate}
\keyword{nonparametric}
\keyword{robust}

