% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationScaleDepth.R
\name{plot,LSDepthContour,ANY-method}
\alias{plot,LSDepthContour,ANY-method}
\title{Plot Location-Scale depth contours.}
\usage{
\S4method{plot}{LSDepthContour,ANY}(
  x,
  cont = NULL,
  ratio = 1,
  mu_min = NULL,
  mu_max = NULL,
  col = NULL,
  border = NULL,
  ...
)
}
\arguments{
\item{x}{object of class LSDepthContour}

\item{cont}{plotted contours. Default NULL means that all contours stored in x will be plotted.}

\item{ratio}{ratio}

\item{mu_min}{mu_min}

\item{mu_max}{mu_max}

\item{col}{vectors with area colors passed to polygon function}

\item{border}{vector with colors for borders}

\item{...}{other parameters passed to polygon}
}
\description{
Create location-scale depth plot. See \code{\link{lsdSampleDepthContours}} for more information.
}
\examples{

smp <- rf(100, 5, 10)
x <- DepthProc::lsdSampleDepthContours(smp)
plot(x, col = paste0("grey", col = rev(seq(10, 40, 10))))

}
