% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot_depthproc}
\alias{plot_depthproc}
\title{plot}
\usage{
\S4method{plot}{DepthCurve,missing}(x, y = "missing")
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
}
\description{
Plot Depth curve
}

