% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RobReg-class}
\alias{RobReg-class}
\title{RobReg}
\description{
Virtual class for robust regression methods from depthproc package
}
\section{Slots}{

\describe{
\item{\code{coef}}{coefficients of fitted model}
}}

