% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleDepthContours}
\alias{sampleDepthContours}
\title{Calculate sample location scale depth contours}
\usage{
sampleDepthContours(x, depth = c(0.1, 0.2, 0.3, 0.4), lengthmu = 1000)
}
\arguments{
\item{x}{one dimensional vector with sample}

\item{depth}{depth level for contours}

\item{lengthmu}{number of points to evalute depth}
}
\description{
Calculate sample location scale depth contours
}
\examples{
## Example for F-distribution
dcont = sampleDepthContours(rf(200,4,7))
plot(dcont)

## Example for normal distribution
## - more contours calculated
dcont_norm = sampleDepthContours(rnorm(100),seq(0.05,0.4,0.05))
plot(dcont_norm)
}

