% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthproc.R
\docType{data}
\name{USLABOUR}
\alias{USLABOUR}
\title{US Labour dataset}
\format{
A data frame with 654 rows and 4 variables
}
\source{
U.S.Department of Labor --- Bureau of Labour Statistics FRED
}
\usage{
data(USLABOUR)
}
\description{
US Labour dataset
}
\keyword{datasets}
