% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{abline,RobReg-method}
\alias{abline,RobReg-method}
\title{Add line to plot}
\usage{
\S4method{abline}{RobReg}(a = NULL, b = NULL, h = NULL, v = NULL,
  reg = NULL, coef = NULL, untf = FALSE, ...)
}
\arguments{
\item{a}{an object of class RobReg}

\item{b}{not used.}

\item{h}{not supported.}

\item{v}{not supported.}

\item{reg}{not supported.}

\item{coef}{not supported.}

\item{untf}{not supported.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
Add fitted line to a plot. This is overloaded function for robust regression methods from package depthproc.
}
