% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,DepthCurveList}
\alias{as.matrix,DepthCurveList-method}
\title{as.matrix method for DepthCurveList.}
\usage{
as.matrix(x, ...)

\S4method{as.matrix}{DepthCurveList}(x)
}
\arguments{
\item{x}{an object of class that inherits from DepthCurveList (ScaleCurveList or AsymmetryCurveList).}

\item{...}{other arguments passed to standard as.matrix function.}
}
\description{
Create a matrix from DepthCurve and DepthCurveList.
}

