% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{AsymmetryCurve-class}
\alias{AsymmetryCurve-class}
\title{AsymmetryCurve and AsymmetryCurveList}
\description{
AsymmetryCurve is a class that stores results of \link{asymmetryCurve} function.
}
\details{
The mechanism of creating plots with multiple curves is shown in \link{DepthCurve-class} (same mechanism is applied for ScaleCurve).
}

