% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BinnDepth2d-class}
\alias{BinnDepth2d-class}
\title{BinnDepth2d}
\description{
Class that stores result of function binningDepth2D(...)
}
\section{Slots}{

\describe{
\item{\code{freq}}{Matrix with number of elements in certain bin.}

\item{\code{mid_x}}{Middle values on x-axis.}

\item{\code{mid_y}}{Middle values on y-axis.}

\item{\code{breaks_x}}{Boundaries of bins.}

\item{\code{breaks_y}}{Boundaries of bins.}

\item{\code{input_data}}{Binned data.}

\item{\code{max_depth_x}}{Point with maximum depth on x-axis.}

\item{\code{max_depth_y}}{Point with maximum depth on y-axis.}
}}

