\docType{class}
\name{BinnDepth2d}
\alias{BinnDepth2d}
\title{BinnDepth2d}
\description{
Class that stores result of function binningDepth2D(...)
}
\section{Slots}{
  \describe{ \item{freq}{Matrix with number of elements in
  certain bin.} \item{mid_x}{Middle values on x-axis.}
  \item{mid_y}{Middle values on y-axis.}
  \item{breaks_x}{Boundaries of bins.}
  \item{breaks_y}{Boundaries of bins.}
  \item{input_data}{Binned data.} \item{max_depth_x}{Point
  with maximum depth on x-axis.} \item{max_depth_y}{Point
  with maximum depth on y-axis.} }
}

