\name{binningDepth2D}
\alias{binningDepth2D}
\title{2d Binning}
\usage{
binningDepth2D(x, nbins = 8, remove_borders = FALSE, devel = FALSE)
}
\description{
Binning 2d
}
\examples{
binningDepth2D(x = mvrnorm(100,rep(0,2),diag(2)))
}
\seealso{
\code{\link{depth}}
}
\keyword{depth}
\keyword{function}
\keyword{multivariate}
\keyword{nonparametric}
\keyword{robust}

