% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DDPlot-class}
\alias{DDPlot-class}
\title{DDPlot}
\description{
Class fro DDPlot
}
\section{Slots}{

\describe{
\item{\code{X}}{Object of class \link{Depth-class}.}

\item{\code{Y}}{Object of class \link{Depth-class}.}

\item{\code{title}}{title of a plot.}

\item{\code{name}}{name of the x data set, used to label the horizontal axis.}

\item{\code{name_y}}{name of the y data set, used to label the vertical axis.}
}}

