% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DeepReg2d-class}
\alias{DeepReg2d-class}
\title{DeepReg2d}
\description{
Class for robust regression methods from depthproc package
}
\section{Slots}{

\describe{
\item{\code{coef}}{coefficients of fitted model}

\item{\code{depth}}{regression depth of the fitted values}
}}

