% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deepReg2d}
\alias{deepReg2d}
\title{Simple deepest regression method.}
\usage{
deepReg2d(x, y)
}
\arguments{
\item{x}{Independent variable.}

\item{y}{Dependent variable.}
}
\description{
This function calculate estimator of simple regression model y=ax+b
}
\details{
Function returns object of class DeepReg2d.
}
\examples{
data(pension)
plot(pension)
abline(lm(Reserves~Income,data = pension), lty = 3, lwd = 2) #lm
abline(deepReg2d(pension[,1],pension[,2]), lwd = 2) #deepreg2d
}
\author{
Daniel Kosiorowski, Mateusz Bocian, Anna Wegrzynkiewicz and Zygmunt Zawadzki from Cracow University of Economics.
}
\references{
Rousseeuw J.P., Hubert M. (1998), Regression Depth, \emph{Journal of The American Statistical Association},  vol.94.
}

