% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Depth-class}
\alias{Depth-class}
\title{Depth}
\description{
Virtual class with structure for every depth class from depthproc package.
}
\section{Slots}{

\describe{
\item{\code{u}}{data set.}

\item{\code{X}}{reference set.}

\item{\code{method}}{depth type.}
}}

