% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Depth-class}
\alias{Depth-class}
\alias{DepthEuclid}
\alias{DepthLP}
\alias{DepthLocal}
\alias{DepthMahalanobis}
\alias{DepthProjection}
\alias{DepthTukey}
\title{Depth}
\description{
Virtual class with structure for every depth class from depthproc package.
}
\section{Slots}{

\describe{
\item{\code{Matrix}}{with number of elements in certain bin.}

\item{\code{X}}{Middle values on x-axis.}

\item{\code{method}}{Middle values on y-axis.}

\item{\code{name}}{Boundaries of bins.}
}}

