% Generated by roxygen2 (4.0.1): do not edit by hand
\name{depthLocal}
\alias{depthLocal}
\title{Local depth}
\usage{
depthLocal(u, X, beta = 0.5, depth1 = "Projection", depth2 = depth1,
  name = "X", ...)
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data as a matrix, data frame. If it is a matrix or data frame, then each row is viewed as one multivariate observation.}

\item{beta}{cutoff value for neighbourhood}

\item{depth1}{depth method for symmetrised data}

\item{depth2}{depth method for calculation depth of given point}

\item{...}{additional parameters passed to depth1 and depth2}
}
\description{
Local depth
}
\examples{
require(MASS)
 data = mvrnorm(100, c(0,5), diag(2)*5)
 #by default depth2 = depth1
 depthLocal(data, data, depth1 = "LP")
 depthLocal(data, data, depth1 = "LP", depth2 = "Projection")
 ## Depthcontour
 depthContour(data, method = "Local", depth1 = "LP")
}

