% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{depthMedian}
\alias{depthMedian}
\alias{depthMedian,Depth}
\alias{depthMedian,Depth-method}
\alias{depthMedian,matrix}
\alias{depthMedian,matrix-method}
\title{Depth median}
\usage{
depthMedian(x, ...)

\S4method{depthMedian}{matrix}(x, ...)

\S4method{depthMedian}{Depth}(x)
}
\arguments{
\item{x}{object of class Depth or matrix.}

\item{\dots}{arguments passed to \code{\link{depth}} function (e.g method).}

\item{...}{arguments passed to \code{\link{depth}} function (e.g method).}
}
\description{
Return point with maximum depth function value. If multiple points have the same value, mean average of them will be returned.
}

