% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fncBoxPlot}
\alias{fncBoxPlot}
\title{Functional boxplot based on Modified Band Depth}
\usage{
fncBoxPlot(u, X = NULL, bands = c(0, 0.5), method = "MBD",
  type = "ggplot2", ...)
}
\arguments{
\item{u}{data matrix}

\item{X}{reference set. If null u will be used as reference.}

\item{bands}{limits for bands}

\item{method}{depth method}

\item{type}{method used for create plot ("ggplot2" or "base")}

\item{\dots}{other arguments passed to fncDepth}
}
\description{
Functional boxplot based on Modified Band Depth
}
\examples{
# some data:
x  = matrix(rnorm(200), ncol = 10)

fncBoxPlot(x, bands = c(0, 0.5, 1), method = "FM")

# fncBoxPlot handles zoo and xts objects
library(xts)
time = as.POSIXct(1:ncol(x) * 86400, origin = "1970-01-01")
x_xts = xts(t(x), order.by = time)
fncBoxPlot(x_xts, bands = c(0, 0.5, 1), method = "FM")
}

