% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_boxplot.R
\name{fncBoxPlot}
\alias{fncBoxPlot}
\title{Functional boxplot based on Modified Band Depth}
\usage{
fncBoxPlot(u, X = NULL, bands = c(0, 0.5), method = "MBD", byrow = NULL, ...)
}
\arguments{
\item{u}{data matrix}

\item{X}{reference set. If null u will be used as reference.}

\item{bands}{limits for bands}

\item{method}{depth method}

\item{byrow}{byrow}

\item{\dots}{other arguments passed to fncDepth}
}
\description{
Functional boxplot based on Modified Band Depth
}
\examples{

# some data:
x <- matrix(rnorm(200), ncol = 10)

DepthProc::fncBoxPlot(x, bands = c(0, 0.5, 1), method = "MBD")
DepthProc::fncBoxPlot(x, bands = c(0, 0.5, 1), method = "MBD", byrow = FALSE)

colnames(x) <- paste0("f", 1:ncol(x))
DepthProc::fncBoxPlot(x, bands = c(0, 0.5, 1), method = "MBD")

# fncBoxPlot handles zoo and xts objects
x <- matrix(rnorm(200), ncol = 10)
time <- as.POSIXct(1:ncol(x) * 86400, origin = "1970-01-01")
x_xts <- xts::xts(t(x), order.by = time)
DepthProc::fncBoxPlot(x_xts, bands = c(0, 0.5, 1), method = "FM")

data("katowice.airpollution", package = "DepthProc")
pl <- DepthProc::fncBoxPlot(katowice.airpollution, bands = c(0, 0.5, 1), method = "MBD")
pl + ggplot2::ggtitle("Air pollution in Katowice") +
  ggplot2::labs(y = "pollination ", x = "hour ")

}
