% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fncDepth}
\alias{fncDepth}
\alias{fncDepth.matrix}
\alias{fncDepth.zoo}
\title{Basic function for functional depths}
\usage{
fncDepth(u, X = NULL, method = "MBD", name = deparse(substitute(u)), ...)

\method{fncDepth}{matrix}(u, X = NULL, method = "MBD",
  name = deparse(substitute(u)), ...)

\method{fncDepth}{zoo}(u, X, method = "MBD", name = deparse(substitute(u)),
  ...)
}
\description{
Calculates depth functions.
}
\examples{
x = matrix(rnorm(60), nc = 20)
fncDepth(x, method = "FM", dep1d = "Mahalanobis")
fncDepth(x)

library(xts)
data(sample_matrix)
sample.xts <- as.xts(sample_matrix, descr='my new xts object')
sample.xts
}

