% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fncDepthFM}
\alias{fncDepthFM}
\title{FM Depth}
\usage{
fncDepthFM(u, X, dep1d = "Projection", ...)
}
\arguments{
\item{u}{Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.}

\item{X}{The data as a matrix. If it is a matrix or data frame, then each row is viewed as one multivariate observation.}

\item{dep1d}{method for depth used in 1d.}

\item{\dots}{other arguments passed to depth function.}
}
\description{
Computes Frainman-Muniz depth for functional data.
}
\examples{
x = matrix(rnorm(60), nc = 20)
fncDepthFM(x)
}

