% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FunctionalDepth-class}
\alias{DepthMBD}
\alias{FunctionalDepth-class}
\title{Functional Depth}
\description{
Virtual class with structure for every functional depth class from depthproc package.
}
\section{Slots}{

\describe{
\item{\code{u}}{data set.}

\item{\code{X}}{reference set.}

\item{\code{method}}{functional depth type.}

\item{\code{name}}{name that will be used on plots. By default it is a name of variable passed to fncDepth.}
}}

