% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getLSDContour,LSDepthContour-method}
\alias{getLSDContour,LSDepthContour-method}
\title{Get location scale contour from LSDepthContour object}
\usage{
\S4method{getLSDContour}{LSDepthContour}(x, cont, ...)
}
\arguments{
\item{x}{object of class LSDepthContour}

\item{cont}{single numeric - depth of contour to return}
}
\description{
Get location scale contour from LSDepthContour object
}
\examples{
dcont = sampleDepthContours(rf(200,4,7), depth = c(0.1,0.2))

#get contour that is present in dcont object
getLSDContour(dcont,0.1)

# get contour that is not present in dcont
# it will be automatically calculated
getLSDContour(dcont,0.3)
}

