% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AsymmetryCurveClass.R,
%   R/DDPlotClass.R, R/ScaleCurveClass.R
\docType{methods}
\name{getPlot}
\alias{getPlot}
\alias{getPlot,AsymmetryCurveList-method}
\alias{getPlot,DDPlot-method}
\alias{getPlot,ScaleCurveList-method}
\title{Create ggplot object from DepthCurve, DepthCurveList and DDPlot classes.}
\usage{
getPlot(object)

\S4method{getPlot}{AsymmetryCurveList}(object)

\S4method{getPlot}{DDPlot}(object)

\S4method{getPlot}{ScaleCurveList}(object)
}
\arguments{
\item{object}{a DDPlot ScaleCurve or AsymmetryCurve object class.}
}
\description{
Create an object of class ggplot from DepthCurve and DepthCurveList.
}
