% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{\%+\%}
\alias{\%+\%}
\alias{\%+\%,DepthCurve,DepthCurve-method}
\alias{\%+\%,DepthCurve,DepthCurveList-method}
\alias{\%+\%,DepthCurveList,DepthCurve-method}
\alias{grapes-plus-grapes-methods,DepthCurve,DepthCurve}
\alias{grapes-plus-grapes-methods,DepthCurve,DepthCurveList}
\alias{grapes-plus-grapes-methods,DepthCurveList,DepthCurve}
\title{Adds plots}
\usage{
e1 \%+\% e2

\S4method{\%+\%}{DepthCurveList,DepthCurve}(e1, e2)

\S4method{\%+\%}{DepthCurve,DepthCurveList}(e1, e2)

\S4method{\%+\%}{DepthCurve,DepthCurve}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{object}
}
\description{
Adds plots
}
\details{
See \code{\link{DepthCurve}} for description.
}

