% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DepthCurveClass.R
\docType{methods}
\name{\%+\%}
\alias{DepthCurve}
\alias{DepthCurveList}
\alias{\%+\%}
\alias{\%+\%,DepthCurve,DepthCurve-method}
\alias{\%+\%,DepthCurve,DepthCurveList-method}
\alias{\%+\%,DepthCurveList,DepthCurve-method}
\alias{grapes-plus-grapes-methods}
\title{Adds plots}
\usage{
e1 \%+\% e2

\S4method{\%+\%}{DepthCurveList,DepthCurve}(e1, e2)

\S4method{\%+\%}{DepthCurve,DepthCurveList}(e1, e2)

\S4method{\%+\%}{DepthCurve,DepthCurve}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{object}
}
\description{
Adds plots
}
\details{
See \code{\link{DepthCurve-class}} for description.
}

