% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{lsdGetContour}
\alias{lsdGetContour}
\alias{lsdGetContour,LSDepthContour}
\alias{lsdGetContour,LSDepthContour-method}
\title{Get location-scale contour from LSDepthContour object}
\usage{
lsdGetContour(x, cont)

\S4method{lsdGetContour}{LSDepthContour}(x, cont)
}
\arguments{
\item{x}{object of class LSDepthContour}

\item{cont}{single numeric - depth of contour to return}
}
\description{
Get location-scale contour from LSDepthContour object
}
\examples{
dcont = lsdSampleDepthContours(rf(200,4,7), depth = c(0.1,0.2))

#get contour that is present in dcont object
lsdGetContour(dcont,0.1)

# get contour that is not present in dcont
# it will be automatically calculated
lsdGetContour(dcont,0.3)
}

