% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mWilcoxonTest}
\alias{mWilcoxonTest}
\title{Depth based multivariate Wilcoxon test for scale difference}
\usage{
mWilcoxonTest(x, y, alternative = "greater")
}
\description{
Depth based multivariate Wilcoxon test for scale difference
}
\examples{
require(MASS)
x = mvrnorm(100, c(0,0), diag(2))
y = mvrnorm(100, c(0,0), diag(2)*1.4)
mWilcoxonTest(x,y)
}

