% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,BinnDepth2d,ANY-method}
\alias{plot,BinnDepth2d,ANY-method}
\title{2d Binning plot}
\usage{
\S4method{plot}{BinnDepth2d,ANY}(x, ..., alpha = 0.1, bg_col = "red",
  add_mid = TRUE)
}
\arguments{
\item{x}{object of class BinnDepth2d}

\item{...}{graphical parameters passed to plot}

\item{alpha}{alpha value for rgb function}

\item{bg_col}{backgroud color}

\item{add_mid}{logical. If TRUE centers of binns will be marked.}
}
\description{
Binning 2d
}
\examples{
require(MASS)
 tmp = binningDepth2D(x = mvrnorm(100,rep(0,2),diag(2)))
 plot(tmp)
}
\seealso{
\code{\link{depth}}
}
\keyword{depth}
\keyword{function}
\keyword{multivariate}
\keyword{nonparametric}
\keyword{robust}

