% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleDepthContours}
\alias{sampleDepthContours}
\title{Calculate sample location scale depth contours}
\usage{
sampleDepthContours(x, depth = c(0.1, 0.2, 0.3, 0.4), lengthmu = 1000)
}
\description{
Calculate sample location scale depth contours
}

